rename model classes
