+++ api/src/feeds/impl/datasets_api_impl.py
@@ -17,7 +17,7 @@ from feeds.impl.error_handling import (
     raise_http_error,
     dataset_not_found,
 )
-from feeds.impl.models.gtfs_dataset_api_impl import GtfsDatasetImpl
+from feeds.impl.models.gtfs_dataset_impl import GtfsDatasetImpl
 from feeds_gen.apis.datasets_api_base import BaseDatasetsApi
 from feeds_gen.models.gtfs_dataset import GtfsDataset
