+++ api/src/feeds/impl/models/gtfs_dataset_impl.py
@@ -5,8 +5,8 @@ from packaging.version import Version
 from sqlalchemy.orm import Mapped
 
 from database_gen.sqlacodegen_models import Gtfsdataset, Validationreport
-from feeds.impl.models.bounding_box_api_impl import BoundingBoxImpl
-from feeds.impl.models.validation_report_api_impl import ValidationReportImpl
+from feeds.impl.models.bounding_box_impl import BoundingBoxImpl
+from feeds.impl.models.validation_report_impl import ValidationReportImpl
 from feeds_gen.models.gtfs_dataset import GtfsDataset
